# docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is an API for my Social_Go",
		"title": "Social_Go API",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"name": "API Support",
			"url": "http://www.swagger.io/support",
			"email": "support@swagger.io"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"basePath": "/v1",
	"paths": {
		"/posts": {
			"post": {
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"description": "Creates a post",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["posts"],
				"summary": "Creates a post",
				"parameters": [
					{
						"description": "Post payload",
						"name": "payload",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/main.postField"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Created",
						"schema": {
							"$ref": "#/definitions/main.postField"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {}
					}
				}
			}
		},
		"/posts/{id}": {
			"get": {
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"description": "Fetches a post by ID",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["posts"],
				"summary": "Fetches a post",
				"parameters": [
					{
						"type": "integer",
						"description": "Post ID",
						"name": "id",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/main.postField"
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {}
					}
				}
			},
			"delete": {
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"description": "Delete a post by ID",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["posts"],
				"summary": "Deletes a post",
				"parameters": [
					{
						"type": "integer",
						"description": "Post ID",
						"name": "id",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"204": {
						"description": "No Content",
						"schema": {
							"type": "string"
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {}
					}
				}
			}
		}
	},
	"definitions": {
		"main.postField": {
			"type": "object",
			"required": ["content", "title"],
			"properties": {
				"content": {
					"type": "string",
					"maxLength": 1000
				},
				"tags": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"title": {
					"type": "string",
					"maxLength": 100
				},
				"user_id": {
					"type": "integer"
				}
			}
		}
	},
	"securityDefinitions": {
		"ApiKeyAuth": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	}
}